requirements.txt
aiogram>=3.0,<4.0
pyrogram
tgcrypto
openai-whisper
python-dotenv
soundfile
